backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
appdirs==1.4.4
beautifulsoup4==4.14.2
bs4==0.0.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==46.0.3
cssselect==1.3.0
curl_cffi==0.13.0
deprecation==2.1.0
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
fake-useragent==2.2.0
fastapi==0.119.0
feedparser==6.0.12
frozendict==2.4.6
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
jiter==0.11.1
lxml==6.0.2
multidict==6.7.0
multitasking==0.0.12
numpy==2.3.4
openai==2.5.0
packaging==25.0
pandas==2.3.3
parse==1.20.2
peewee==3.18.2
platformdirs==4.5.0
postgrest==2.22.0
propcache==0.4.1
protobuf==6.33.0
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
pyee==13.0.0
PyJWT==2.10.1
pyppeteer==0.0.25
pyquery==2.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
realtime==2.22.0
requests==2.32.5
requests-html==0.10.0
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
starlette==0.48.0
storage3==2.22.0
StrEnum==0.4.15
supabase==2.22.0
supabase-auth==2.22.0
supabase-functions==2.22.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
w3lib==2.3.1
websockets==15.0.1
yarl==1.22.0
yfinance==0.2.66
